refactor: move hooks and dynamic fonctions
