+++ app/_components/ui/CallToAction/CallToAction.tsx
@@ -1,8 +1,9 @@
 'use client';
-import dynamic from 'next/dynamic';
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { Button } from '@/lib/components/ui/button';
 import CardProps from '@/types/CardProps';
+import { dynamicDownload } from '@/utils/dynamicDownload';
+import { IconLoader } from '@/hooks/IconLoader';
 
 export const CallToAction: React.FC<CardProps> = ({
   cta1,
@@ -14,28 +15,10 @@ export const CallToAction: React.FC<CardProps> = ({
   href2,
   downloadActive2,
 }) => {
-  const downloadAction = (downloadActive: boolean) => {
-    return downloadActive ? { download: true } : { target: '_blank' };
-  };
-
-  const IconLoader = (icon: string) => {
-    const [Icon, setIcon] = useState<React.ComponentType | null>(null);
-
-    useEffect(() => {
-      const loadIcon = async () => {
-        const iconModule = await import(`lucide-react`);
-        setIcon(() => iconModule[icon]);
-      };
-      loadIcon();
-    }, [icon]);
-
-    return Icon ? <Icon className='mr-2' /> : null;
-  };
-
   return (
     <div className='flex h-auto w-full items-center'>
       <div className='mr-4'>
-        <a href={`/${href1}`} {...downloadAction(downloadActive1)}>
+        <a href={`/${href1}`} {...dynamicDownload(downloadActive1)}>
           <Button
             size='lg'
             variant='default'
@@ -47,7 +30,7 @@ export const CallToAction: React.FC<CardProps> = ({
         </a>
       </div>
       <div>
-        <a href={`/${href2}`} {...downloadAction(downloadActive2)}>
+        <a href={`/${href2}`} {...dynamicDownload(downloadActive2)}>
           <Button
             size='lg'
             variant='outline'
