+++ app/types/LocationEventsProps.tsx
@@ -1,0 +1,4 @@
+export interface LocationEventsMapProps {
+  position: [number, number];
+  ref?: React.RefObject<HTMLDivElement>;
+}
