+++ app/utils/constants/baseUrl.tsx
@@ -1,0 +1,1 @@
+export const baseUrl = 'https://';
