+++ app/utils/dynamicDownload.tsx
@@ -1,0 +1,3 @@
+export const dynamicDownload = (downloadActive: boolean) => {
+  return downloadActive ? { download: true } : { target: '_blank' };
+};
