+++ app/utils/dynamicMarginBottom.tsx
@@ -1,0 +1,3 @@
+export const dynamicMarginBottom = (index: number) => {
+  return index < 2 ? 'mb-4' : 'mb-0';
+};
