+++ app/utils/dynamicYear.tsx
@@ -1,0 +1,3 @@
+const year = new Date().getFullYear();
+
+export default year;
